python: Return a list of ISAs from get_runtime_isas.
The get_runtime_isas function is used in two contexts. The first is that
it lets a component specify that it requires support for a certain ISA
in order to work correctly. The second is a less direct way of checking
what the current targetted ISA is.

This change reworks get_runtime_isa to be get_runtime_isas which returns
a list of supported ISAs instead of a single ISA. This is still not
great because it depends on there being a single, universal list of what
ISAs exist, but at least it will give reasonable answers if there are
multiple ISAs configured.

The other use of get_runtime_isa, checking to see what type of ISA is
being simulated as a proxy for what type of ISA and what type of
components are being hooked up, is not as fixed by this change. That
assumption is still present, and will fall apart if multiple ISAs are
configured in, since the function will just tell you if an ISA is
supported, not if a particular system is using it.

Change-Id: I1fee85cc4297d8cb778b1fe5e951d9a4091fe6ca